pyproject.toml
[project]
name = "nhp-model"
dynamic = ["version"]
description = "New Hospital Programme demand model"

requires-python = ">=3.11,<3.14"

dependencies = [
    "azure-identity>=1.12",
    "azure-storage-blob>=12.15",
    "azure-storage-file-datalake>=12.10",
    "importlib-resources>=6.5.2",
    "jsonschema>=4.23.0",
    "numpy>=1.23",
    "pandas>=1.5",
    "pyarrow>=20.0",
    "pyjanitor>=0.23",
    "python-dotenv>=1.0",
    "scipy>=1.10",
    "tqdm>=4.65",
]


# Explicitly specify which packages to include
[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"nhp.model.data.reference" = ["*.csv", "*.json"]
"nhp.model.params" = ["*.json"]

[project.optional-dependencies]
dev = [
    "coverage>=7.2",
    "ipykernel>=6.21",
    "ipython>=8.11",
    "ipywidgets>=8.0",
    "jupyter_client>=8.0",
    "jupyter_core>=5.2",
    "jupyterlab_pygments>=0.2",
    "jupyterlab_widgets>=3.0",
    "matplotlib>=3.7",
    "nbconvert>=7.2",
    "nbformat>=5.7",
    "notebook>=6.5",
    "pygam>=0.8",
    "pytest>=7.2",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "ruff>=0.11.10",
    "setuptools-scm>=8.3.1",
    "snakeviz>=2.1",
    "widgetsnbextension>=4.0",
]
databricks = [
    "pyspark",
    "databricks-connect"
]
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-section-index"
]

[build-system]
requires = ["setuptools>=80", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py311"
indent-width = 4
exclude = ["notebooks"]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = [
    "D", # pydocstyle
    "PLC0415", # `import` should be at the top-level of a file
    "PLR2004", # Magic value used in comparison
    "PD901", # Avoid using the generic variable name `df`
    "RUF005", # list concatenation
]

[tool.ruff.lint]
# Simple rules: pylint + isort
select = [
    "D",   # pydocstyle
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "I",   # isort (import sorting)
    "PL",  # pylint rules,
    "PD",  # pandas-vet
    "NPY", # NumPy-specific rules
    "RUF", # Ruff-specific rules
]

[tool.ruff.lint.isort]
# isort configuration
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 7

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.setuptools_scm]
version_file = "src/nhp/model/_version.py"
fallback_version = "0.0.0"

[tool.ty.src]
exclude = ["notebooks", "docs"]
